1.) Added logger in all auth microservice. 2.) Added Error Logging in every file 3.) Replaced int64 with int 4.) Added error handling where it was missing. 5.) Created a new logging interface and abstracted out the implementation.

+++ auth/entities/user.go
@@ -7,7 +7,7 @@ import (
 )
 
 type User struct {
-	Id       int64  `db:"id"`
+	Id       int    `db:"id"`
 	Username string `db:"username" validate:"required"`
 	Password string `db:"password" validate:"passwd"`
 }
